backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
sqlalchemy>=1.4,<3.0
pydantic>=2.7,<3.0
langgraph>=1.0,<2.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
